requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2023.10
cryptography==45.0.3
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==23.5
django-modeltranslation==0.18.11
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-i18n==0.3.9
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
whitenoise==6.6.0
